v457 - :CustomTable, divide by zero error in thermal code [BENCHMARKED]
Support for custom tabular output
-Added new class CCustomTable and members (CustomOutput.h/.cpp)
-CModel::AddCustomTable (Model.cpp/.h, ParseInput.cpp,

bug fix: divide by zero errors in EnergyTransport.cpp with zero length stream or very fast travel times

minor bug fixes
-minor fixes to DemandOptimization.cpp code
-minor fix to Diagnostics.cpp - return before deletion fixed
-long to int casting in Groundwater code
-minor bug fix in ParseLib.cpp - flagged by unix compiler

+++ src/Drain.cpp
@@ -253,7 +253,7 @@ void CGWDrain::UpdateRatesOfChange(const double      *state_vars,
       if (old_head < _aElevations[index]) { old_head = _aElevations[index]; }
 
       //-- Get HRU-node overlap weight & node area
-      weight = pGWModel->GetOverlapWeight(HRUid, n);
+      weight = pGWModel->GetOverlapWeight((int)(HRUid), n);
       area = pGWModel->GetNodeArea(n);
 
       //-- Calc drain flow change
